Refactoring backup/restore and Makefile

+++ bin/backup.sh
@@ -25,7 +25,7 @@ case "$1" in
         fi
 
         logMsg "Starting MySQL backup..."
-        mysqldump --opt --single-transaction --events --all-databases --routines --comments | bzip2 > "${BACKUP_DIR}/${BACKUP_MYSQL_FILE}"
+        dockerExec mysqldump --opt --single-transaction --events --all-databases --routines --comments | bzip2 > "${BACKUP_DIR}/${BACKUP_MYSQL_FILE}"
         logMsg "Finished"
         ;;
 
@@ -33,13 +33,20 @@ case "$1" in
     ## Solr
     ###################################
     "solr")
-        if [ -f "${BACKUP_DIR}/${BACKUP_SOLR_FILE}" ]; then
-            logMsg "Removing old backup file..."
-            rm -f -- "${BACKUP_DIR}/${BACKUP_SOLR_FILE}"
-        fi
+        if [[ -n "$(dockerContainerId solr)" ]]; then
+            logMsg "Starting Solr backup..."
+            docker-compose stop solr
 
-        logMsg "Starting Solr backup..."
-        tar jcPf "${BACKUP_DIR}/${BACKUP_SOLR_FILE}" /storage/solr/
-        logMsg "Finished"
+            if [ -f "${BACKUP_DIR}/${BACKUP_SOLR_FILE}" ]; then
+                logMsg "Removing old backup file..."
+                rm -f -- "${BACKUP_DIR}/${BACKUP_SOLR_FILE}"
+            fi
+            dockerExec tar -cP --to-stdout /storage/solr/ | bzip2 > "${BACKUP_DIR}/${BACKUP_SOLR_FILE}"
+
+            docker-compose start solr
+            logMsg "Finished"
+        else
+            echo "[WARNING] Skipping solr backup, no such container"
+        fi
         ;;
 esac
